requirements.txt
# Resume/Job Matching System - Dependencies (GPU Optimized)

# Core Framework
flask>=3.0.0
gunicorn>=21.2.0


# NLP & ML (GPU versions installed separately)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Document Parsing
pdfplumber>=0.10.0
python-docx>=1.0.0
PyPDF2>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
fpdf2>=2.7.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Note: PyTorch with CUDA is installed separately via:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
